Overally refactored: remove unncessary parts with rearrangement
