--- bt_xml/default_bt.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<root BTCPP_format="4">
-  <BehaviorTree ID="space-simulator">
-    <Sequence>
-      <LocalSensingNode/>
-      <Fallback>
-        <Sequence>
-          <DecisionMakingNode/>
-          <TaskExecutingNode/>
-        </Sequence>
-        <ExplorationNode/>
-      </Fallback>
-    </Sequence>
-  </BehaviorTree>
-
-  <!-- Description of Node Models (used by Groot) -->
-  <TreeNodesModel>
-    <Action ID="DecisionMakingNode"
-            editable="true"/>
-    <Action ID="ExplorationNode"
-            editable="true"/>
-    <Action ID="LocalSensingNode"
-            editable="true"/>
-    <Action ID="TaskExecutingNode"
-            editable="true"/>
-  </TreeNodesModel>
-
-</root>
