+++ config.yaml
@@ -1,5 +1,7 @@
 scenario: 
-  environment: scenarios.simple.env
+  # environment: scenarios.simple
+  # environment: scenarios.harbor_logistics
+  environment: scenarios.drone_delivery
 
 # decision_making: # Case 1
 #   plugin: plugins.grape.grape.GRAPE
@@ -27,15 +29,15 @@ decision_making: # Case 2
 
 agents:
   behavior_tree_xml: default_bt.xml 
-  quantity: 10
+  quantity: 1
   locations:
     x_min: 0
     x_max: 1400
     y_min: 0
     y_max: 1000
     non_overlap_radius: 0 
-  max_speed: 0.25  
-  max_accel: 0.05
+  max_speed: 3.0  
+  max_accel: 0.5
   max_angular_speed: 0.25
   target_approaching_radius: 50
   work_rate: 1  # work rate for each agent (per 1.0/simulation.sampling_freq)
@@ -44,7 +46,7 @@ agents:
   random_exploration_duration: 1000.0 # sec
 
 tasks:
-  quantity: 100
+  quantity: 20
   locations:
     x_min: 0
     x_max: 1400
@@ -57,9 +59,9 @@ tasks:
     max: 60.0      
   dynamic_task_generation:
     enabled: True
-    interval_seconds: 2000
+    interval_seconds: 500
     max_generations: 3
-    tasks_per_generation: 25
+    tasks_per_generation: 20
 
 simulation:
   sampling_freq: 1.0 
@@ -80,7 +82,7 @@ simulation:
     agent_work_done: True
     agent_assigned_task_id: True
     agent_path_to_assigned_tasks: True
-    task_id: False
+    task_id: True
   saving_options:
     output_folder: output
     with_date_subfolder: True
