+++ config/default/drone_delivery.yaml
@@ -36,7 +36,7 @@ agents:
     y_min: 0
     y_max: 1000
     non_overlap_radius: 0 
-  max_speed: 3.0  
+  max_speed: 1.5  
   max_accel: 0.5
   max_angular_speed: 0.25
   target_approaching_radius: 50
