+++ modules/base_agent.py
@@ -39,7 +39,11 @@ class BaseAgent:
         self.messages_received = []
 
         self.distance_moved = 0.0
-        
+        self.task_amount_done = 0.0
+
+        self.assigned_task_id = None         # Local decision-making result.
+        self.planned_tasks = []              # Local decision-making result.
+
 
     def create_behavior_tree(self, behavior_tree_xml):        
         xml_root = parse_behavior_tree(behavior_tree_xml)        
@@ -168,12 +172,45 @@ class BaseAgent:
         self.update_color()
         pygame.draw.polygon(screen, self.color, [p1, p2, p3])
 
+    def draw_path_to_assigned_tasks(self, screen):
+        # Starting position is the agent's current position
+        start_pos = self.position
+
+        # Define line thickness
+        line_thickness = 3  # Set the desired thickness for the lines        
+        # line_thickness = 16-4*self.agent_id  # Set the desired thickness for the lines        
+
+        # For Debug
+        color_list = [
+            (255, 0, 0),  # Red
+            (0, 255, 0),  # Green
+            (0, 0, 255),  # Blue
+            (255, 255, 0),  # Yellow
+            (255, 0, 255),  # Magenta
+            (0, 255, 255),  # Cyan
+            (255, 165, 0),  # Orange
+            (128, 0, 128),  # Purple
+            (255, 192, 203) # Pink
+        ]
+                
+        # Iterate over the assigned tasks and draw lines connecting them
+        for task in self.planned_tasks:
+            task_position = task.position
+            pygame.draw.line(
+                screen,
+                # (255, 0, 0),  # Color for the path line (Red)
+                color_list[self.agent_id%len(color_list)], 
+                (int(start_pos.x), int(start_pos.y)),
+                (int(task_position.x), int(task_position.y)),
+                line_thickness  # Thickness of the line
+            )
+            # Update the start position for the next segment
+            start_pos = task_position
 
     def draw_tail(self, screen):
         # Draw track
         if len(self.memory_location) >= 2:
             pygame.draw.lines(screen, self.color, False, self.memory_location, 1)               
-        
 
     def draw_communication_topology(self, screen, agents):
      # Draw lines to neighbor agents
@@ -187,11 +224,32 @@ class BaseAgent:
         text_surface = self.font.render(f"agent_id: {self.agent_id}", True, (50, 50, 50))
         screen.blit(text_surface, (self.position[0] + 10, self.position[1] - 10))
 
+    def draw_assigned_task_id(self, screen):
+        # Draw assigned_task_id next to agent position
+        if len(self.planned_tasks) > 0:
+            assigned_task_id_list = [task.task_id for task in self.planned_tasks]
+        else:
+            assigned_task_id_list = self.assigned_task_id
+        text_surface = self.font.render(f"task_id: {assigned_task_id_list}", True, (50, 50, 50))
+        screen.blit(text_surface, (self.position[0] + 10, self.position[1]))
+
+    def draw_work_done(self, screen):
+        # Draw assigned_task_id next to agent position
+        text_surface = self.font.render(f"dist: {self.distance_moved:.1f}", True, (50, 50, 50))
+        screen.blit(text_surface, (self.position[0] + 10, self.position[1] + 10))
+        text_surface = self.font.render(f"work: {self.task_amount_done:.1f}", True, (50, 50, 50))
+        screen.blit(text_surface, (self.position[0] + 10, self.position[1] + 20))
+
     def draw_situation_awareness_circle(self, screen):
         # Draw the situation awareness radius circle    
         if self.situation_awareness_radius > 0:    
             pygame.draw.circle(screen, self.color, (self.position[0], self.position[1]), self.situation_awareness_radius, 1)
 
+    def set_assigned_task_id(self, task_id):
+        self.assigned_task_id = task_id
+
+    def set_planned_tasks(self, task_list): # This is for visualisation
+        self.planned_tasks = task_list    
 
     def set_global_info_agents(self, agents_info):
         self.agents_info = agents_info
@@ -242,3 +300,6 @@ class BaseAgent:
         """ Retrieve tasks that are not yet assigned, specifically where assigned_to is None. """
         unassigned_tasks = [task for task in self.tasks_info if task.assigned_to is None]
         return unassigned_tasks
+
+    def update_task_amount_done(self, amount):
+        self.task_amount_done += amount
