+++ scenarios/simple/task.py
@@ -9,8 +9,7 @@ task_colors = generate_task_colors(config['tasks']['quantity'] + tasks_per_gener
 
 from modules.base_task import BaseTask
 
-sampling_freq = config['simulation']['sampling_freq']
-sampling_time = 1.0 / sampling_freq  # in seconds
+
 class Task(BaseTask):
     def __init__(self, task_id, position):
         super().__init__(task_id, position)
@@ -18,11 +17,6 @@ class Task(BaseTask):
         self.radius = self.amount / config['simulation']['task_visualisation_factor']
         self.color = task_colors.get(self.task_id, (0, 0, 0))  # Default to black if task_id not found
 
-
-    def reduce_amount(self, work_rate):
-        self.amount -= work_rate * sampling_time
-        if self.amount <= 0:
-            self.set_done()
 
     def draw(self, screen):
         self.radius = self.amount / config['simulation']['task_visualisation_factor']        
@@ -34,6 +28,7 @@ class Task(BaseTask):
             font = pygame.font.Font(None, 15)
             text_surface = font.render(f"task_id {self.task_id}: {self.amount:.2f}", True, (250, 250, 250))
             screen.blit(text_surface, (self.position[0], self.position[1]))
+
 
 def generate_tasks(task_quantity=None, task_id_start = 0):
     if task_quantity is None:
